Restrict access with controller actions

+++ app/controllers/articles_controller.rb
@@ -24,10 +24,15 @@ class ArticlesController < ApplicationController
     end  
   end  
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted."
-      redirect_to articles_path 
-    end
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article."
+      redirect_to root_path 
+    else  
+      if @article.destroy
+        flash[:success] = "Article has been deleted."
+        redirect_to articles_path 
+      end
+    end  
   end
   
   def show
